Test: Github Fake 토큰 응답 핸들러 추가
