+++ src/test/java/nextstep/utils/GithubFakeController.java
@@ -1,0 +1,32 @@
+package nextstep.utils;
+
+import lombok.RequiredArgsConstructor;
+import nextstep.auth.token.oauth2.github.GithubAccessTokenRequest;
+import nextstep.auth.token.oauth2.github.GithubAccessTokenResponse;
+import nextstep.auth.token.oauth2.github.GithubClient;
+import nextstep.auth.token.oauth2.github.GithubProfileResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class GithubFakeController {
+
+    @PostMapping("/github/login/oauth/access_token")
+    public ResponseEntity<GithubAccessTokenResponse> accessToken(@RequestBody GithubAccessTokenRequest request) {
+        String accessToken = GithubFakeResponse.getAccessToken(request.getCode());
+
+        return ResponseEntity.ok().body(GithubAccessTokenResponse.builder()
+            .accessToken(accessToken)
+            .build());
+    }
+
+    @GetMapping("/github/user")
+    public ResponseEntity<GithubProfileResponse> user() {
+        return ResponseEntity.ok().build();
+    }
+}
+
